Move LegacyBoostrapUpdatesFinder.kt to a package
