+++ plugins/core/src/main/kotlin/de/fayard/refreshVersions/core/internal/legacy/LegacyBoostrapUpdatesFinder.kt
@@ -1,8 +1,13 @@
-package de.fayard.refreshVersions.core.internal
+package de.fayard.refreshVersions.core.internal.legacy
 
 import de.fayard.refreshVersions.core.ModuleId
 import de.fayard.refreshVersions.core.RefreshVersionsCorePlugin
 import de.fayard.refreshVersions.core.Version
+import de.fayard.refreshVersions.core.internal.DependencyWithVersionCandidates
+import de.fayard.refreshVersions.core.internal.RefreshVersionsConfigHolder
+import de.fayard.refreshVersions.core.internal.VersionCandidatesResultMode
+import de.fayard.refreshVersions.core.internal.getDependencyVersionFetchers
+import de.fayard.refreshVersions.core.internal.getVersionCandidates
 import okhttp3.OkHttpClient
 
 internal object LegacyBoostrapUpdatesFinder {
